rename channel pruning learners
