+++ learners/learner_utils.py
@@ -21,8 +21,8 @@ import tensorflow as tf
 from learners.full_precision.learner import FullPrecLearner
 from learners.weight_sparsification.learner import WeightSparseLearner
 from learners.channel_pruning.learner import ChannelPrunedLearner
-from learners.channel_pruning_legacy.learner import ChannelPrunedLegacyLearner
 from learners.channel_pruning_gpu.learner import ChannelPrunedGpuLearner
+from learners.channel_pruning_rmt.learner import ChannelPrunedRmtLearner
 from learners.discr_channel_pruning.learner import DisChnPrunedLearner
 from learners.uniform_quantization.learner import UniformQuantLearner
 from learners.uniform_quantization_tf.learner import UniformQuantTFLearner
@@ -48,10 +48,10 @@ def create_learner(sm_writer, model_helper):
     learner = WeightSparseLearner(sm_writer, model_helper)
   elif FLAGS.learner == 'chn-pruned':
     learner = ChannelPrunedLearner(sm_writer, model_helper)
-  elif FLAGS.learner == 'chn-pruned-lgc':
-    learner = ChannelPrunedLegacyLearner(sm_writer, model_helper)
   elif FLAGS.learner == 'chn-pruned-gpu':
     learner = ChannelPrunedGpuLearner(sm_writer, model_helper)
+  elif FLAGS.learner == 'chn-pruned-rmt':
+    learner = ChannelPrunedRmtLearner(sm_writer, model_helper)
   elif FLAGS.learner == 'dis-chn-pruned':
     learner = DisChnPrunedLearner(sm_writer, model_helper)
   elif FLAGS.learner == 'uniform':
